Amdocs-genai/backend/requirements.txt
# Web framework for backend APIs
Flask==2.2.5

# Parsing and handling JSON responses
Flask-Cors==3.0.10

# Image processing for image analysis
Pillow==9.4.0
opencv-python-headless==4.5.5.64

# Hugging Face Transformers for NLP and deepfake detection
transformers==4.30.2
datasets==2.14.4

# Fact-checking, sentiment analysis, and other NLP tasks
torch==2.0.1
sentencepiece==0.1.99

# Deepfake detection
timm==0.9.2

# Web scraping utilities
beautifulsoup4==4.12.2
requests==2.31.0

# Database driver
sqlite3-binary==3.41.0.post0

# Utilities for file handling and performance optimization
numpy==1.25.2
scikit-learn==1.3.0

# Test libraries
pytest==7.4.2
pytest-mock==3.11.1

# Frontend testing for React components
pytest-django==4.8.0
